The big refactor of the connection handler to use a single ZK thread.
Uses a deque and pipe to signal the ZK thread that there's a new command to send, so that the ZK thread can send it, or retrieve a response.
Processing ZK requests and responses serially in a single thread eliminates the need for a bunch of the locking, and two threads working on the same underlying socket.
There's still an intermittent test failure or two, probably related to the complete removal of the state lock which will need some slight re-ordering to remedy.