make Close and Ping print the same as the rest
This makes the log message look like

Sending request: Ping()

instead of

Sending request: <class 'kazoo.protocol.serialization.Ping'>

Constructing a Ping object for each ping generates a lot of unnecessary
temporary objects, though, so this also creates a singleton object to
use for it.

It makes the equivalent change to Close for consistency's sake.