sphinx: Configure sys.path to allow autodoc to find Kazoo.
Previously, if Kazoo wasn't already in sys.path otherwise (e.g. if not
installed) sphinx autodoc wouldn't be able to find the module for
documenting purposes because sphinx-build happens in an adjacent
subdirectory.

This adds the docs folder's parent to the path so that autodoc can find
the Kazoo module properly and do its stuff.