Handle network partition in SetPartitioner receipe
If connection gets suspended, change state to ALLOCATING if we had already ACQUIRED.
This way the caller does not process the members since we could eventually lose session
and get repartitioned. If we got connected after a suspension it means we've not lost
the session and still have our members. Hence, restore to ACQUIRED