Don't treat null (or None) data as ""
In ZK's Java impl if the data is null, it'll be serialized
with length equals -1. If it is an empty string, it'll be serialized
with length equals 0.

Currently, Kazoo treats data with a length of -1 as "". This diff
changes it so that it'll come back as None. Same thing for creating
and setting data to None. Tests added.

Reported-by: Grant Monroe <tnarg@twitter.com>
Signed-off-by: Raul Gutierrez S <rgs@itevenworks.net>