Add a new public method remove_watches()
This implementation covers up ZOOKEEPER-442. I went
with a simpler API than the one provided by the Java
implementation. If there's a need to remove individual
watchers, I think that should be provided via a separate
method (i.e.: unregister watchers) and it should be
the caller's responsability to issue a separate
remove_watches() call to remove the watch from ZK.

Signed-off-by: Raul Gutierrez S <rgs@itevenworks.net>