Add retries to server_version() fetch
Due to the apparent flakiness of the four-letter commands
we need to add a set of retries to the fetching of the
server version to handle getting partial or truncated values
back from it.

This adds a default of 4 tries (1 + 3 retries) to the server
version code so that if a partial value is returned (or no
server version at all) that the code will try again, if none
can be fetched in 4 tries; then none is returned to signify
this failure.

Fixes issue #274