Implement read-write (shared) lock from ZooKeeper recipe.
Reference: http://zookeeper.apache.org/doc/trunk/recipes.html#Shared+Locks

This is a writer-preference shared lock.

Most of the lock protocol is the same as for an exclusive lock, except that it needs to be parameterized on:
1) Node name - lock "kind" - a write lock (__lock__) or read lock (__rlock__).
2) Which node names block which other node names (read locks block only writers, writers block both).
3) Finding the predecessor node according to the rule in #2.

These changes are integrated to the main Lock class. The changes are backward compatible with existing users of Lock. Users should normally never pass the Lock node_name and exclude_names kwargs directly, they should use WLock and RLock.