Ensure zookeeper is cleanly torn down after each test
Instead of keeping a zookeeper around that has been
previously used, at each test boundary start the zookeeper
cluster and at the end, stop each server and reset each
server so that the tests each see a clean zookeeper cluster.

This also adds checks to make sure that on each stop that
is called that the zookeeper process has exited cleanly with
a zero error code.