Fix semaphore usage
The semaphore should be passing the blocking argument to the
underlying lock that is used (so it is non-blocking as well) and
it should also be passing the timeout as well.

The non-blocking result should also not add a watch that will
later attempt to get the least (another call to acquire should be
used instead, or the blocking mode should be used instead).