Handle system interrupts during socket.create_connection
This is in a similar vein to #250, but the root cause of this issue is actually an upstream python issue (http://bugs.python.org/issue20611) which is not fixed in 2.x or <3.5. Python doesn't handle interrupted system calls-- so the applications are responsible for doing so.

This patch simply retries the create_connection call on system call interrupt, while honoring the original timeout request.

Note: although the timeout for `create_tcp_connection` will be honored, if there are interrupts each subsequent call to `create_connection` will have less time to complete.