Set watch on immediate predecessor in locks
The original lock code would set the watch on the last node which
preceded our node.  This is efficient because if there are N
predecessors, then we would only set the watch on the latest and
would therefore only wake up and go through the loop after all
the predecessors had released their locks.  The change to support
shared locks, however, set the watch on the first node which
preceded ours, meaning that we will wake up and iterate through
the loop for all N predecessor nodes.

This patch restores the original behavior where the watch was only
set on the immediate predecessor.