Correct LockingQueue to work with multiple readers.
The current LockingQueue implementation implicitly assumes (in check_for_updates()) that once it has a list of available and locked queue items, the state of the first available but not taken item will not change before it can take the item. This is not true in the presence of multiple reader processes. This has two negative consequences:

1. If the item is processed and removed by another process before the item is taken, the taken entry is created and remains until the connection is closed. If connections are long-lasting, taken entries for processed and removed queue items accumulate.
2. If the item is either taken or processed and removed by another process, the call to _inner_get() will block until there is more activity on the queue, even if there are more items available for processing. Reader processes will be blocked unnecessarily.

So, remove the assumption. If a take fails, consider other items in the queue before blocking. And tidy taken entries if, after creation, the entry is found to be no longer on the queue.