feat(core): add support for GSSAPI authentication
This commit adds the possibility to connect to Zookeeper using
the GSSAPI mechanism (Kerberos). A new scheme is added (gssapi)
which requires to precise a Kerberos principal or None if the
local credential cache should be used instead. Moreover, the
kerberos_server_principal (kdc) must also be set when
initializing the KazooClient object. Documentation is updated.