fix(core): ensure timeout argument is positive
Previously, a gap between calls to `time.time()` could lead to a
situation where the current time was less than `end` during the
`while` condition, but it was greater than `end` when assigning
a value to `timeout_at`.

Add tests to ensure a socket.error is raised instead of passing a
nonpositive value as a timeout to socket.create_connection.