WIP feat(core) only fail queued requests if we are not CONNECTING (#374)
If the connection is lost, but the state is 'CONNECTING', the client
is trying to revalidate an existing session.

When that happens, the requests which have been dequeued but are still
pending are interrupted with a 'ConnectionLoss' exception--as we don't
know if the packet reached the server, and the ACKs will never come
back anyway.

Without this patch, the requests which are still queued, and thus
haven't been emitted are also cancelled with the same exception
(see bottom half of '_notify_pending').

It seems that there is no reason to cancel such requests when the new
state is 'CONNECTING', as the client is trying to validate an existing
session; these "pristine" requests could very well be submitted
in-order over the new connection if the session is recovered.

This also seems more consistent with the first patch for issue #374:
it does not matter whether a request was queued just before or during
the "outage"--as long as it was not emitted.

This patch implements that feature.

ATTN: The patch is marked WIP because it should most probably *NOT* be
merged--as it turns out that both the Java and the C client cancel
such requests when the connection is lost.

Java: ClientCnxn socket error -> cleanAndNotifyState -> cleanup
    -> conLossPacket/remove

C: Socket error -> handle_socket_error_msg -> handle_error -> cleanup
    -> cleanup_bufs -> free_buffers/free_completions