recipe(lock): retry lock cleanup
If a non-blocking lock fails to acquire the lock, and then encounters
a KazooError (due to a suspended session), the _best_effort_cleanup method
will swallow the exception and control will return without the lock
contender node being deleted.  If the session resumes (does not expire)
then we will have left a lock contender in place, which will eventually
become an orphaned, stuck lock once the original actor releases it.

To correct this, retry deleting the lock contender in all cases.  Due
to the importance of this, we ignore the supplied timeout (in case the
aquire method was called with a timeout) and retry forever.

Closes: #732